ChargePoint simulator: Implement shed logic
A logic for shedding a port load in implemented, according to
the CP API description. These simualtors APIs are affected:
 ShedLoad
 ClearShedState

ShedLoad API sheds the load of a single port, all ports of a station or
all stations in a group. The absolute load can be specified or % of the
current load. A duration for this sheding can be configured.

ClearShedState clears all load shedding on a single staton or on all
stations from a group.

Signed-off-by: Tzvetomir Stoyanov (VMware) <tz.stoyanov@gmail.com>