ChargePoint simulator: Add support for fixed CP infrastructure for the random simulator
When the simulator works in random mode, it generates randomly the whole
charging infrastructure. However, there are use cases where this charging
infrastructure should not be randomly generated each time the simulator
starts.
Support for fixed charging infrastructure is added, defined in
the configuration file. A few configuration files with predefined
facilities are added.

Signed-off-by: Tzvetomir Stoyanov (VMware) <tz.stoyanov@gmail.com>